Index and styles was separeted, fix routes and text input styles
