+++ Reservacion/src/components/Check/styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+export const styles=StyleSheet.create({
+    containerCheck:{
+        flexDirection: 'row',
+        alignItems: 'center',
+        color: '#b6b7ba',
+    },
+})
