+++ Reservacion/src/components/InputText/styles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    input: {
+      height: 40,
+      margin: 12,
+      borderBottomWidth: 3,
+      padding: 10,
+    },
+})
