+++ Reservacion/src/components/TextPrimary/styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    text:{
+        fontWeight: 'bold',
+        marginVertical: 4,
+        marginHorizontal: 0
+    }
+})
