+++ Reservacion/src/components/TextSecond/styles.js
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    text:{
+        marginVertical: 4,
+        marginHorizontal: 0
+    }
+})
