+++ Reservacion/src/screens/BookingScreen/index.js
@@ -1,20 +1,17 @@
-import React,{useState} from 'react';
-import { View, StyleSheet } from 'react-native';
-import { ButtonPrimary } from '../components/ButtonPrimary/index';
+import React from 'react';
+import { View,  } from 'react-native';
+import { ButtonPrimary } from '../../components/ButtonPrimary/index';
 import Icon from 'react-native-vector-icons/FontAwesome5';
-import { TextPrimary } from '../components/TextPrimary/index';
-import { InputText } from '../components/InputText';
+import { TextPrimary } from '../../components/TextPrimary';
+import { InputText } from '../../components/InputText';
+import { styles } from './styles';
 
 export const BookingScreen = ({navegation}) => {
-
-
-
 
   return (
     <View style={styles.container}>
          <Icon name="angle-left" size={25} color="#5b6ef7"/>
-        
-        
+                
         <View style={styles.containerText}> 
         <TextPrimary 
             text={'Where are you now'} 
@@ -33,14 +30,3 @@ export const BookingScreen = ({navegation}) => {
   );
 };
 
-const styles = StyleSheet.create({
-    container:{
-      backgroundColor: 'white',
-      height: '100%',
-      paddingHorizontal: 22,
-    },
-    containerButton:{
-      pasition: 'absolute',
-      top:'40%',
-    }
-})
