+++ Reservacion/src/screens/BookingScreen/styles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    container:{
+      backgroundColor: 'white',
+      height: '100%',
+      paddingHorizontal: 22,
+    },
+    containerButton:{
+      pasition: 'absolute',
+      top:'40%',
+    }
+})
