+++ Reservacion/src/screens/HomeScreeen/styles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+    container:{
+      backgroundColor: 'white',
+      height: '100%',
+      padding: 15,
+    },
+    content:{
+      pasition: 'absolute',
+      top:'28%',
+    }
+})
