--- Reservacion/src/screens/MyFlightsScreen.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { View, StyleSheet, TextInput } from 'react-native';
-import { ButtonPrimary } from '../components/ButtonPrimary/index';
-import { TextPrimary } from '../components/TextPrimary/index';
-import { TextSecond } from '../components/TextSecond';
-
-export const MyFlightsScreen = ({navigation}) => {
-    
-
-
-
-
-
-};
-
