+++ Reservacion/src/screens/MyFlightsScreen/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { View, TextInput } from 'react-native';
+import { ButtonPrimary } from '../../components/ButtonPrimary/index';
+import { TextPrimary } from '../../components/TextPrimary/index';
+import { TextSecond } from '../../components/TextSecond';
+
+export const MyFlightsScreen = ({navigation}) => {
+    
+
+
+
+
+
+};
