+++ Reservacion/src/screens/MyFlightsScreen/styles.js
@@ -1,0 +1,1 @@
+import { StyleSheet } from 'react-native';
